+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [] -
+## [3.1] - 2022-01-13
 ### Changed
 - Speed-up CI tests by caching installation of libs and splitting tests into randomized groups using pytest-test-groups
 - Mock Ensembl services (gene conversions and liftover) used all tests